Please enter the commit message for your changes. Lines starting with '#' will be ignored, and an empty message aborts the commit.
On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	modified:   logic_building/functions/recursion/1_reverse_number.exe
	modified:   logic_building/functions/recursion/2_reverse_number_v2.exe
	modified:   logic_building/functions/recursion/3_reverse_number_v3.c
	modified:   logic_building/functions/recursion/3_reverse_number_v3.exe

+++ logic_building/functions/recursion/3_reverse_number_v3.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 
-int helper_reverse(int n , int rev){
+size_t helper_reverse(size_t n , size_t rev){
 
     if(n == 0 ){
 
@@ -12,16 +12,16 @@ int helper_reverse(int n , int rev){
 
 }
 
-int reverse(int n){
+size_t reverse(size_t n){
 
     return helper_reverse(n,0);
 
 }
 
-int main(void){
+size_t main(void){
 
-    int n = 12345;
-    printf("%d",reverse(n));
+    size_t n = 12345;
+    printf("%zu",reverse(n));
 
 
 }
